clskj

+++ ft_print.h
@@ -9,7 +9,7 @@
 # include <stdarg.h>
 
 int 	ft_printf(const char *fmt_or_nch, ...);
-size_t		ft_putnbr_string(int subject_num, const char *base_char);
+int		ft_putnbr_string(int subject_num, const char *base_char);
 size_t		ft_putchar(char inputchar_num);
 size_t	ft_putchar_string(char *sub_string);
 char	*ft_strdup(const char *s);
